archived_cdn_api_async/src/requirements.txt
redis==4.4.4
fastapi==0.104.0
orjson==3.9.9
pydantic==1.9.1
uvicorn==0.23.2
python-dotenv==1.0
gunicorn==21.2.0
uvloop==0.18.0 ; sys_platform != "win32" and implementation_name == "cpython"
httptools==0.6.1
aiohttp==3.8.6
boto3==1.28.68
aioboto3==11.3.1
aiobotocore==2.7.0
miniopy_async==1.16
geopy==2.4.0
aiofiles==23.2.1
aioshutil==1.3
python-magic
APScheduler==3.10.4